+++ admissions/tests/test_utils.py
@@ -202,13 +202,15 @@ class TestObfuscateAdmission(TestCase):
         self.alex.refresh_from_db()
         self.sander.refresh_from_db()
 
-        self.assertNotEqual(self.alex.first_name, "Alexander")
-        self.assertNotEqual(self.alex.last_name, "Orvik")
+        self.assertNotEqual(
+            self.alex.first_name + self.alex.last_name, "AlexanderOrvik"
+        )
         self.assertNotEqual(self.alex.phone, "12345678")
         self.assertNotEqual(self.alex.address, "Klostergata 35")
 
-        self.assertNotEqual(self.sander.first_name, "Sander")
-        self.assertNotEqual(self.sander.last_name, "Haga")
+        self.assertNotEqual(
+            self.sander.first_name + self.sander.last_name, "SanderHaga"
+        )
         self.assertNotEqual(self.sander.phone, "87654321")
         self.assertNotEqual(self.sander.address, "Klostergata 35")
 
@@ -220,7 +222,7 @@ class TestInterviewGenerationEdgeCases(TestCase):
         self.bodegaen = InterviewLocation.objects.create(name="Bodegaen")
 
         # Initialize the start of the interview period to 12:00
-        self.start = datetime.date.today()
+        self.start = timezone.datetime.today()
         self.datetime_start = date_time_combiner(self.start, datetime.time(hour=12))
 
         # End of interview period is two days later giving us a three day interview period
@@ -232,11 +234,13 @@ class TestInterviewGenerationEdgeCases(TestCase):
             default_interview_day_start=datetime.time(hour=12),
             default_interview_day_end=datetime.time(hour=20),
         )
+
         InterviewLocationAvailability.objects.create(
             interview_location=self.knaus,
             datetime_from=self.datetime_start,
             datetime_to=self.datetime_start + timezone.timedelta(hours=8),
         )
+
         InterviewLocationAvailability.objects.create(
             interview_location=self.bodegaen,
             datetime_from=self.datetime_start + timezone.timedelta(hours=4),